use "size" of the repo to weight quality metrics
The GitHub repository API returns a `size`, which is (allegedly) in KB
units. We'll use this size to compute a "weighted mean" of our GPA and
code coverage metrics from CodeClimate.

Side-note: we're hardcoding the `size` for echo-chat repo because our
code is a very small share of that repository and we don't want it to
be weighted too heavily.